backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
requests
starlette
sqlalchemy
PyJWT
psycopg2-binary
bcrypt